+++ src/com/perdi/backend/group/Group.java
@@ -1,14 +1,20 @@
 package com.perdi.backend.group;
-import java.util.ArrayList;
 
+import java.util.ArrayList;
+import java.util.Random;
+
+import com.perdi.backend.userpkg.User;
 import com.perdi.backend.postpkg.Post;
 
-import User;
-
 public class Group {
+    private int id;
     private ArrayList<User> members;
     private ArrayList<Post> posts;
     //Usuario vai ser a classe criada
+    public Group() {
+        Random rand = new Random()
+        this.id = rand.nextInt(0,999999);
+    }
     public void putUser(User new_member) {
         members.add(new_member);
     }
@@ -32,5 +38,8 @@ public class Group {
     }
     public int getUserCount() {
         return this.members.size();
+    }
+    public int getId() {
+        return id;
     }
 }
